// src/containers/layout/Layout.enum.ts
export enum LayoutAlignmentVertical {
	TOP = 'top',
	CENTER = 'center',
	BOTTOM = 'bottom'
}

export enum LayoutAlignmentHorizontal {
	LEFT = 'left',
	CENTER = 'center',
	RIGHT = 'right'
}

export enum LayoutFlow {
	OVERLAP = 'overlap',
	ROW = 'row',
	COLUMN = 'column'
}

export enum LayoutFit {
	WIDTH = 'width',
	HEIGHT = 'height',
	BOTH = 'both'
}